backend/requirements.txt
# ==================== Web框架 ====================
fastapi>=0.100.0
uvicorn>=0.20.0

# ==================== 数据库 ====================
SQLAlchemy>=2.0.0
pymysql>=1.0.2

# ==================== 缓存与存储 ====================
redis>=5.0.0
minio>=7.1.0

# ==================== 认证与安全 ====================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator

# ==================== AI与HTTP请求 ====================
openai
requests>=2.28.0

# ==================== 任务调度 ====================
apscheduler>=3.10.0

# ==================== 工具库 ====================
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# ==================== 开发工具 ====================
ruff>=0.1.6
pre-commit>=3.5.0
